Add Exercise 28_03
Note my solution for the nonrecursive DFS corrects an error from the homework's algorithm. The true search path should be the following if choosing ascending neighbors:
Chicago Seattle San Francisco Los Angeles Denver Kansas City New York Boston Atlanta Miami Houston Dallas

The homework has the difference of going to Denver then Los Angeles but this does not make sense when considering from the perspective of San Francisco. From San Francisco, Los Angeles is the next edge in the list (since to Seattle is not considered).

The error from the homework's algorithm is in the for loop. The edges are being deleted which changes the size of neighbors.get(v).size() but the counter i continues going up. This ends up skipping Los Angeles and choosing Denver (even though both are unvisisted).